compose-working--master/api/requirements-windows-local.txt
cffi
Django
djangorestframework
django-environ
django-filter
django-autoslug
django-countries
django-phonenumber-field
phonenumbers
psycopg2
gunicorn
django-cors-headers
django-debug-toolbar
djoser
djangorestframework_simplejwt
Fabric3
django-rest-swagger
django-templated-email
celery
redis
flower
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
pytest
pytest-django
model_bakery
pytest-watch
isort
flake8
autopep8
locust
channels
channels-redis
stripe
fpdf2
exponent_server_sdk
rsa
cryptography
